+++ src/SchoolAccount.Infrastructure/Helpers/Filtering/Filters/SubTaskFilterableRegistrar.cs
@@ -25,7 +25,7 @@ public class SubTaskFilterableRegistrar : IFilterableRegistrar<CalendarOfItemsFi
         {
             [Keys.Name] = (Expression<Func<SubTaskEntity, string>>)(x => x.Name),
             [Keys.Categories] =
-                (Expression<Func<SubTaskEntity, IEnumerable<int>>>)(x => x.Task.TypeTaskMappings.Select(t => t.TypeId)),
+                (Expression<Func<SubTaskEntity, IEnumerable<int>>>)(x => x.Task.TypeTaskMappings.Select(t => (int)t.TypeId)),
             [Keys.State] = (Expression<Func<SubTaskEntity, WorkflowState>>)(x => x.WorkflowState),
             [Keys.PhaseOfEducation] =
                 (Expression<Func<SubTaskEntity, IEnumerable<long>>>)(x => x.TagsSourceMappings.Select(t => t.TagId)),
